pyproject.toml
[project]
name = "nana-wiki"
version = "0.0.1"
description = "Your sovereign, multilingual AI knowledge assistant for Obsidian"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Nana Project" }]
requires-python = ">=3.12"
keywords = ["obsidian", "knowledge-base", "ai", "wiki", "cli", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "pyyaml>=6.0",
    "questionary>=2.0.0",
    "prompt_toolkit>=3.0.0",
    "requests>=2.31.0",
    "pillow>=10.0.0",
    "rich-pixels>=3.0.0",
]

[project.scripts]
nana = "nana.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nana"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
